+++ src/helpers/changelogData.js
@@ -1,1 +1,1 @@
-export const appVersion = "0.6.0 (Prototype)";
+export const appVersion = "v0.8.0 (Prototype)";